+++ src/macros.rs
@@ -1,21 +1,5 @@
 //! Macros shared throughout the compiler-builtins implementation
 
-/// Changes the visibility to `pub` if feature "public-test-deps" is set
-#[cfg(not(feature = "public-test-deps"))]
-macro_rules! public_test_dep {
-    ($(#[$($meta:meta)*])* pub(crate) $ident:ident $($tokens:tt)*) => {
-        $(#[$($meta)*])* pub(crate) $ident $($tokens)*
-    };
-}
-
-/// Changes the visibility to `pub` if feature "public-test-deps" is set
-#[cfg(feature = "public-test-deps")]
-macro_rules! public_test_dep {
-    {$(#[$($meta:meta)*])* pub(crate) $ident:ident $($tokens:tt)*} => {
-        $(#[$($meta)*])* pub $ident $($tokens)*
-    };
-}
-
 /// The "main macro" used for defining intrinsics.
 ///
 /// The compiler-builtins library is super platform-specific with tons of crazy
